% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign_pkgs.R
\name{extract.tobit1}
\alias{extract.tobit1}
\title{Extract function to be used with the \code{texreg} package.}
\usage{
extract.tobit1(model, include.logLik = TRUE, include.nobs = TRUE, ...)
}
\arguments{
\item{model}{An object of class \code{ergmito}.}

\item{include.logLik}{if \code{TRUE} the log-likelihood value is included}

\item{include.nobs}{if \code{TRUE} the number of observations is included}

\item{...}{Further arguments}
}
\description{
This function is inspired by the ERGMitos package
}
