% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{feesadm}
\alias{feesadm}
\title{Fees and admission}
\format{
a tibble containing:
\itemize{
\item fees: expenses on fees and admission,
\item expense: total expense,
\item netinc: net income
}
}
\description{
Expenses on fees and admission from the US expense survey
}
\keyword{dataset}
