% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tobit1.R
\name{hausman}
\alias{hausman}
\title{Hausman test}
\usage{
hausman(x, y, omit = NULL)
}
\arguments{
\item{x}{the first model,}

\item{y}{the second model}

\item{omit}{a character containing the effects that are removed from the test}
}
\description{
Hausman test
}
\author{
Yves Croissant
}
