% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{haustest}
\alias{haustest}
\title{Hausman test}
\usage{
haustest(x, y, omit = NULL)
}
\arguments{
\item{x}{the first model,}

\item{y}{the second model}

\item{omit}{a character containing the effects that are removed from the test}
}
\description{
Hausman test
}
\examples{
library("tidyverse")
charitable <- charitable \%>\% mutate(logdon = log(donation) - log(25))
char_form <- logdon ~ log(donparents) + log(income) +
    education + religion + married + south
ml <- tobit1(char_form, data = charitable)
scls <- update(ml, method = "trimmed")
haustest(scls, ml, omit = "(Intercept)")
}
\references{
\insertRef{HAUS:78}{tobit1}
}
\author{
Yves Croissant
}
