% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{moffitt}
\alias{moffitt}
\title{Moffitt data set}
\format{
a tibble containing:
\itemize{
\item hours: hours worked per week,
\item wage: hourly wage,
\item nwinc: asset income per week,
\item married: dummy for married women,
\item age: age of the women,
\item black: dummy for black women,
\item fsize: size of the family
\item clt6: number of children under 6 years old,
\item cgt6: number of children above 6 years old,
\item educ: number of years of education,
\item lfsize: size of the labor force (in millions),
\item manuf: manufacturing fraction,
\item gov: government fraction.
}
}
\source{
this data set was kindly provided by David Drukker
}
\description{
a sample of 610 women drawn from the 1972 wave of the
national Longitudinal Survey (NLS) of Older Women
}
\details{
Labor force participation and hours worked by women
}
\references{
\insertRef{MOFF:84}{tobit1}
\insertRef{SKEE:VELL:99}{tobit1}
\insertRef{DRUK:02}{tobit1}
}
\keyword{dataset}
