% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign_pkgs.R
\name{prediction.tobit1}
\alias{prediction.tobit1}
\title{prediction method}
\usage{
\method{prediction}{tobit1}(
  model,
  data = find_data(model, parent.frame()),
  at = NULL,
  what = c("expvalue", "prob", "linpred"),
  ...
)
}
\arguments{
\item{model}{a model fitted using \code{tobit1}}

\item{data, at}{see \code{prediction::prediction}}

\item{what}{the kind of predictions}

\item{...}{further arguments}
}
\description{
The prediction package computes the predicted values in a tidy way,
using the predict method of the object. It is used by the margins package
}
