% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tobit1.R
\name{tobit1}
\alias{tobit1}
\alias{nobs.tobit1}
\alias{vcov.tobit1}
\alias{logLik.tobit1}
\alias{summary.tobit1}
\alias{print.tobit1}
\alias{print.summary.tobit1}
\title{Truncated response model}
\usage{
tobit1(
  formula,
  data,
  subset = NULL,
  start = NULL,
  left = 0,
  right = Inf,
  sample = c("censored", "truncated"),
  method = c("ml", "lm", "2steps", "trimmed", "nls"),
  trace = FALSE
)

\method{nobs}{tobit1}(object, ...)

\method{vcov}{tobit1}(object, ...)

\method{logLik}{tobit1}(object, ...)

\method{summary}{tobit1}(object, ...)

\method{print}{tobit1}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{print}{summary.tobit1}(
  x,
  digits = max(3, getOption("digits") - 2),
  width = getOption("width"),
  ...
)
}
\arguments{
\item{formula}{a symbolic description of the model,}

\item{data}{a data frame,}

\item{subset}{a subset,}

\item{start}{an optional vector of starting values}

\item{left, right}{left and right limits of the dependent
variable. The default is respectively 0 and +Inf which
corresponds to the most classic (left-zero truncated) tobit
model,}

\item{sample}{either \code{"censored"} (the default) to estimate the
censored (tobit) regression model or \code{"truncated"} to estimated
the truncated regression model,}

\item{method}{one of \code{"ml"} for maximum likelihood, \code{"lm"} for
(biased) least squares estimators and \code{"2steps"} for two-steps
consistent estimators, \code{"trimmed"} for symetrically censored
estimator,}

\item{trace}{a boolean (the default if \code{FALSE}) if \code{TRUE} some
information about the optimization process is printed,}

\item{...}{further arguments.}

\item{x, object}{an object of class \code{tobit1} or \code{summary.tobit1},}

\item{digits, width}{see \code{base::print},}
}
\description{
Estimation of models for which the response is truncated, either on
censored or truncated samples using linear models, maximum
likelihood or two-steps estimators
}
\examples{
# tobit model estimated by maximum likelihood
tobit1(fees ~ expense, feesadm)
# same using two-steps estimator
tobit1(fees ~ expense, feesadm, method = "2steps")
# same model fitted on the truncated sample
tobit1(fees ~ expense, feesadm, sample = "truncated")
}
\author{
Yves Croissant
}
