% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tradeprotection}
\alias{tradeprotection}
\title{Lobying and Trade Protection}
\format{
a tibble containing:
\itemize{
\item ntb NTB coverage ratio, proportion
\item exports exportations
\item imports importations
\item elast demand elasticity
\item cap lobying
\item labvar labor market covariate
\item sic3 3-digit SIC industry classification
\item inv Inventories, factor share
\item engsci Engineers and scientists, factor share
\item whitecol White collar, factor share
\item skill Skilled, factor share
\item semskill Semi-skilled, factor share
\item cropland Cropland, factor shaer
\item pasture Pasture, factor share
\item forest Forest, factor share
\item coal Coal, factor share
\item petro Petroleum, factor share
\item minerals Minerals, factor share
\item scrconc Seller concentration
\item bcrconc Buyer concentration
\item scrcomp Seller number of firms
\item bcrcomp Buyer number of firms
\item meps Scale
\item kstock Capital stock
\item puni bla
\item geog2 Geographic concentration
\item tenure Average worker tenure, years
\item klratio Capital-labor ratio
\item bunion bla
}
}
\source{
American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}
}
\description{
194 industrial sectors in the US
}
\details{
Lobying from Capitalists and Unions and Trade Protection
}
\references{
\insertRef{MATS:SHER:06}{tobit1}
}
\keyword{dataset}
